refactor: shared client and remove Text
