+++ providers/dns/internal/rimuhosting/fixtures/add_record_same_domain.xml
