--- providers/dns/zonomi/internal/fixtures/delete_record_error.xml
@@ -1,2 +1,0 @@
-<?xml version ="1.0"  ?><!DOCTYPE html [<!ENTITY nbsp '&#160;'><!ENTITY trade '&#8482;'><!ENTITY copy '&#169;'>]>
-<error>ERROR: No zone found for example.com</error>
