--- providers/dns/zonomi/internal/fixtures/delete_record_nothing.xml
@@ -1,13 +1,0 @@
-<?xml version ="1.0"  ?><!DOCTYPE html [<!ENTITY nbsp '&#160;'><!ENTITY trade '&#8482;'><!ENTITY copy '&#169;'>]><dnsapi_result><is_ok>OK:</is_ok>
-    <result_counts
-            added="0"
-            changed="0"
-            unchanged="0"
-            deleted="0"/>
-    <actions>
-
-        <action
-                action="DELETE"
-                host="example.org"
-                type="TXT"
-                value="aaaa"></action></actions></dnsapi_result>
