--- providers/dns/zonomi/internal/fixtures/find_records_pattern.xml
@@ -1,24 +1,0 @@
-<?xml version ="1.0"  ?><!DOCTYPE html [<!ENTITY nbsp '&#160;'><!ENTITY trade '&#8482;'><!ENTITY copy '&#169;'>]><dnsapi_result><is_ok>OK:</is_ok>
-    <result_counts
-            added="0"
-            changed="0"
-            unchanged="0"
-            deleted="0"/>
-    <actions>
-
-        <action
-                action="QUERY"
-                host="_test.example.org, example.org"
-                type="TXT">
-            <record
-                    name="_test.example.org"
-                    type="TXT"
-                    content="txttxtx"
-                    ttl="3600 seconds"
-                    prio="0"/>
-            <record
-                    name="example.org"
-                    type="TXT"
-                    content="txttxtx"
-                    ttl="3600 seconds"
-                    prio="0"/>	</action></actions></dnsapi_result>
